wiring/component: convert power wiring data to be floats
Fixes #335.

+++ src/component/power_component.cc
@@ -13,9 +13,9 @@ power_component_manager::create_component_instance_data(unsigned count) {
     instance_data new_pool;
 
     size_t size = sizeof(c_entity) * count;
-    size = sizeof(unsigned) * count + align_size<unsigned>(size);
+    size = sizeof(float) * count + align_size<float>(size);
     size = sizeof(bool) * count + align_size<bool>(size);
-    size = sizeof(unsigned) * count + align_size<unsigned>(size);
+    size = sizeof(float) * count + align_size<float>(size);
     size += 16;   // for worst-case misalignment of initial ptr
 
     new_buffer.buffer = malloc(size);
@@ -24,14 +24,14 @@ power_component_manager::create_component_instance_data(unsigned count) {
     memset(new_buffer.buffer, 0, size);
 
     new_pool.entity = align_ptr((c_entity *)new_buffer.buffer);
-    new_pool.required_power = align_ptr((unsigned *)(new_pool.entity + count));
+    new_pool.required_power = align_ptr((float *)(new_pool.entity + count));
     new_pool.powered = align_ptr((bool *)(new_pool.required_power + count));
-    new_pool.max_required_power = align_ptr((unsigned *)(new_pool.powered + count));
+    new_pool.max_required_power = align_ptr((float *)(new_pool.powered + count));
 
     memcpy(new_pool.entity, instance_pool.entity, buffer.num * sizeof(c_entity));
-    memcpy(new_pool.required_power, instance_pool.required_power, buffer.num * sizeof(unsigned));
+    memcpy(new_pool.required_power, instance_pool.required_power, buffer.num * sizeof(float));
     memcpy(new_pool.powered, instance_pool.powered, buffer.num * sizeof(bool));
-    memcpy(new_pool.max_required_power, instance_pool.max_required_power, buffer.num * sizeof(unsigned));
+    memcpy(new_pool.max_required_power, instance_pool.max_required_power, buffer.num * sizeof(float));
 
     free(buffer.buffer);
     buffer = new_buffer;
