+++ src/wiring/wiring_data.h
@@ -9,9 +9,9 @@ enum wire_type {
 
 
 struct power_wiring_data {
-    unsigned total_power = 0;
-    unsigned total_draw = 0;
-    unsigned peak_draw = 0;
+    float total_power = 0;
+    float total_draw = 0;
+    float peak_draw = 0;
     unsigned num_consumers = 0;
     unsigned num_providers = 0;
 };
